Move core traits and types to tide-core
