+++ Cargo.toml
@@ -2,6 +2,7 @@
 members = [
     "tide",
     "tide-compression",
+    "tide-core",
     "examples",
 ]
