+++ tide-core/src/app.rs
@@ -31,6 +31,7 @@ use crate::{
 ///
 /// ```rust, no_run
 /// #![feature(async_await)]
+/// # use tide_core as tide;
 ///
 /// let mut app = tide::App::new();
 /// app.at("/hello").get(async move |_| "Hello, world!");
@@ -46,6 +47,7 @@ use crate::{
 ///
 /// ```rust, no_run
 /// #![feature(async_await)]
+/// # use tide_core as tide;
 ///
 /// use tide::error::ResultExt;
 ///
@@ -76,6 +78,7 @@ use crate::{
 ///
 /// ```rust, no_run
 /// #![feature(async_await)]
+/// # use tide_core as tide;
 ///
 /// use http::status::StatusCode;
 /// use serde::{Deserialize, Serialize};
@@ -167,7 +170,7 @@ impl<State: Send + Sync + 'static> App<State> {
     ///
     /// ```rust,no_run
     /// # #![feature(async_await)]
-    /// # let mut app = tide::App::new();
+    /// # let mut app = tide_core::App::new();
     /// app.at("/").get(async move |_| "Hello, world!");
     /// ```
     ///
@@ -191,7 +194,7 @@ impl<State: Send + Sync + 'static> App<State> {
     /// Here are some examples omitting the HTTP verb based endpoint selection:
     ///
     /// ```rust,no_run
-    /// # let mut app = tide::App::new();
+    /// # let mut app = tide_core::App::new();
     /// app.at("/");
     /// app.at("/hello");
     /// app.at("add_two/:num");
